Changes proposed by NikRoe implemented
